+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Bug with polyfill for `integer` rule
+- Bug with `Infinity` values in range-based rules
 
 ## [1.2.1] - 2018-08-17